+++ Gateway/Command/FetchTransactionCommand.php
@@ -4,6 +4,8 @@ namespace Dintero\Checkout\Gateway\Command;
 
 use Dintero\Checkout\Model\Api\Client;
 use Magento\Payment\Gateway\CommandInterface;
+use Magento\Framework\Registry;
+use Magento\Sales\Model\Order\Payment\Transaction;
 
 /**
  * Class VoidCommand
@@ -20,20 +22,48 @@ class FetchTransactionCommand implements CommandInterface
     private $api;
 
     /**
+     * @var Registry $registry
+     */
+    private $registry;
+
+    /**
      * Capture constructor.
      *
      * @param Client $client
+     * @param Registry $registry
      */
-    public function __construct(Client $client)
-    {
+    public function __construct(
+        Client $client,
+        \Magento\Framework\Registry $registry
+    ) {
         $this->api = $client;
+        $this->registry = $registry;
     }
 
     /**
-     * Executing command
+     * Updating transaction type
+     *
+     * @param array $responseData
+     * @return void
+     */
+    private function updateTransactionType($responseData)
+    {
+        /** @var Transaction $transaction */
+        $transaction = $this->registry->registry('current_transaction');
+
+        if ($transaction
+            && $transaction->getTxnType() === Transaction::TYPE_ORDER
+            && $responseData['status'] === Client::STATUS_AUTHORIZED
+        ) {
+            $transaction->setTxnType(Transaction::TYPE_AUTH);
+        }
+    }
+
+    /**
+     * Fetching transaction information
      *
      * @param array $commandSubject
-     * @return $this|\Magento\Payment\Gateway\Command\ResultInterface|null
+     * @return array
      * @throws \Magento\Payment\Gateway\Http\ClientException
      * @throws \Magento\Payment\Gateway\Http\ConverterException
      */
@@ -42,7 +72,7 @@ class FetchTransactionCommand implements CommandInterface
         /** @var \Magento\Sales\Model\Order\Payment $payment */
         $payment = $commandSubject['payment']->getPayment();
         $result = $this->api->getTransaction(
-            $commandSubject['transaction_id'],
+            $commandSubject['transactionId'],
             $payment->getOrder()->getStoreId()
         );
 
@@ -50,7 +80,14 @@ class FetchTransactionCommand implements CommandInterface
             throw new \Exception(__('Failed to void the transaction'));
         }
 
-        $payment->setTransactionId($commandSubject['transaction_id']);
-        return $result;
+        $payment->setTransactionId($commandSubject['transactionId']);
+        $this->updateTransactionType($result);
+        $data = [];
+        foreach ($result as $field => $value) {
+            if (is_string($value)) {
+                $data[$field] = $value;
+            }
+        }
+        return $data;
     }
 }