+++ Model/Api/Client.php
@@ -254,7 +254,7 @@ class Client
             'Dintero-System-Name' => __('Magento'),
             'Dintero-System-Version' => $this->getSystemMeta()->getVersion(),
             'Dintero-System-Plugin-Name' => 'Dintero.Checkout.Magento.V2',
-            'Dintero-System-Plugin-Version' => '1.7.8',
+            'Dintero-System-Plugin-Version' => '1.7.9',
         ];
 
         if ($token && $token instanceof Token) {